renames About to Settings and fixes color picker
