+++ ai2svelte/src/js/main/Components/TabBar.svelte
@@ -80,12 +80,12 @@
     <div class="tab-items">
       {@render tab("Home", true)}
       {@render tab("CSS", false)}
-      {@render tab("Preview", false)}
-      {@render tab("About", false)}
+      {@render tab("Settings", false)}
     </div>
     <div class="theme-configs">
       <ThemeSwitcher bind:theme />
       <div
+        id="picker-accent"
         style="--picker-color: {accentColor};"
         use:tooltip={{
           ...tooltipSettings,
